fix(decompilation): use cfr lib instead of procyon
Super types detection was failing in some cases with procyon. Switching
to cfr, which seems to be more active, fixes the issue.

Closes STUDIO-4079
